isolate PostManager
