+++ Sources/Component/Component.swift
@@ -12,6 +12,7 @@ public protocol Component: ViewRenderable {
 
     var contentMode: ComponentContentMode { get set }
 
+    /// A component should render at least once for showing its content.
     /// The rendering should only happen on the main thread.
     func render()
